put all of simplepool into one file (netlist.py). Same for tests.
