+++ assets/netlists/simplepool/test/test_netlist.py
@@ -1,8 +1,11 @@
-"""simply test for scope"""
+from enforce_typing import enforce_types
 import inspect
-from .. import netlist
+import pytest
 
-def test1():
+from .. import netlist
+from util.constants import S_PER_DAY
+
+def test_scope():
     #example usage
     ss = netlist.SimStrategy()
     ss = netlist.SimState()
@@ -13,3 +16,51 @@ def test1():
     assert inspect.isclass(netlist.KPIs)
     assert callable(netlist.netlist_createLogData)
     assert callable(netlist.netlist_plotInstructions)
+
+@enforce_types
+def test_SimStrategy():
+    #import from `netlist` module, not a `SimState` module. Netlist has it all:)
+    ss = netlist.SimStrategy()
+
+    assert ss.pub_init_OCEAN > 0.0
+    assert ss.OCEAN_init > 0.0
+    assert ss.OCEAN_stake > 0.0
+
+    assert ss.pub_init_OCEAN >= (ss.OCEAN_init + ss.OCEAN_stake)
+    
+    assert ss.DT_init > 0.0
+    assert ss.DT_stake > 0.0
+    assert ss.DT_init >= ss.DT_stake
+
+    assert (ss.pool_weight_DT + ss.pool_weight_OCEAN) == 10.0
+
+
+@enforce_types
+def test_SimState():
+    #import from `netlist` module
+    state = netlist.SimState()
+
+    assert len(state.agents) == 1
+    
+    assert state.getAgent("pub1").OCEAN() == state.ss.pub_init_OCEAN
+
+    for i in range(1000):
+        state.takeStep()
+        if len(state.agents) > 1:
+            break
+
+    assert len(state.agents) > 1
+    pool_agents = state.agents.filterToPool()
+    assert len(pool_agents) > 0
+    
+@enforce_types
+def test_KPIs():
+    #import from `netlist` module
+    kpis = netlist.KPIs(time_step=12)
+    
+    assert kpis.tick() == 0
+    kpis.takeStep(state=None)
+    kpis.takeStep(state=None)
+    assert kpis.tick() == 2
+    assert kpis.elapsedTime() == 12*2
+
